pyproject.toml
[project]
name = "chat-etl"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jeonghoon2", email = "ukidd12@naver.com"},
]
dependencies = [
    "kafka-python>=2.0.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
